requirements.txt
amqp==1.4.6
gevent==1.0.2
greenlet==0.4.7
hiredis==0.2.0
PyYAML==3.11
redis==2.10.3
requests==2.7.0
plyvel==0.9
netaddr==0.7.15
antlr4-python2-runtime==4.5.2
jmespath==0.7.1
click==4.1

